🏗️ Separate CLI code from library (Closes #49)
