+++ catalog.go
@@ -1,4 +1,4 @@
-package main
+package ffcss
 
 import (
 	"fmt"
@@ -20,10 +20,10 @@ type Catalog map[string]Theme
 func (store Catalog) Lookup(query string) (Theme, error) {
 	originalQuery := query
 	query = lookupPreprocess(query)
-	d("using query %q", query)
+	D("using query %q", query)
 	processedThemeNames := make([]string, 0, len(store))
 	for _, theme := range store {
-		d("\tlooking up against %q (%q)", lookupPreprocess(theme.Name()), theme.Name())
+		D("\tlooking up against %q (%q)", lookupPreprocess(theme.Name()), theme.Name())
 		if lookupPreprocess(theme.Name()) == query {
 			return theme, nil
 		}
@@ -54,7 +54,7 @@ func LoadCatalog(storeDirectory string) (themes Catalog, err error) {
 	if err != nil {
 		return
 	}
-	d("loading potential themes %v into catalog", func() []string {
+	D("loading potential themes %v into catalog", func() []string {
 		dirNames := make([]string, 0, len(manifests))
 		for _, dir := range manifests {
 			dirNames = append(dirNames, dir.Name())
@@ -70,7 +70,7 @@ func LoadCatalog(storeDirectory string) (themes Catalog, err error) {
 		if err != nil {
 			return nil, err
 		}
-		d("\tadding theme from manifest %q", manifest.Name())
+		D("\tadding theme from manifest %q", manifest.Name())
 		themes[themeName] = theme
 	}
 	return
