+++ cmd/ffcss/get_test.go
@@ -1,0 +1,1 @@
+package main
