+++ currenttheme_test.go
@@ -1,1 +1,1 @@
-package main
+package ffcss
