+++ firefoxversion.go
@@ -1,4 +1,4 @@
-package main
+package ffcss
 
 import (
 	"fmt"
@@ -17,10 +17,10 @@ type FirefoxVersion struct {
 
 // FirefoxVersion represents a constraint to test on a firefox version.
 type FirefoxVersionConstraint struct {
-	min FirefoxVersion
-	max FirefoxVersion
-	// To be included in a sentence like "this theme ensures compatibility with firefox <sentence>"
-	sentence string
+	Min FirefoxVersion
+	Max FirefoxVersion
+	// To be included in a Sentence like "this theme ensures compatibility with firefox <Sentence>"
+	Sentence string
 }
 
 // NewFirefoxVersionConstraint creates a firefox version constraint from its string representation.
@@ -40,14 +40,14 @@ func NewFirefoxVersionConstraint(constraint string) (FirefoxVersionConstraint, e
 	var sentence string
 	var err error
 	if strings.HasSuffix(constraint, "+") {
-		d("constraint type is minimum")
+		D("constraint type is minimum")
 		min, err = NewFirefoxVersion(strings.TrimSuffix(constraint, "+"))
 		if err != nil {
 			return FirefoxVersionConstraint{}, fmt.Errorf("while parsing minimum constraint %q: %w", constraint, err)
 		}
 		sentence = fmt.Sprintf("version %s or higher", min)
 	} else if strings.Count(constraint, "-") == 1 {
-		d("constraint type is range")
+		D("constraint type is range")
 		minmaxStrings := strings.SplitN(constraint, "-", 2)
 		min, err = NewFirefoxVersion(minmaxStrings[0])
 		if err != nil {
@@ -59,14 +59,14 @@ func NewFirefoxVersionConstraint(constraint string) (FirefoxVersionConstraint, e
 		}
 		sentence = min.String() + "–" + max.String()
 	} else if strings.HasPrefix(constraint, "up to ") {
-		d("constraint type is upto")
+		D("constraint type is upto")
 		max, err = NewFirefoxVersion(strings.TrimPrefix(constraint, "up to "))
 		if err != nil {
 			return FirefoxVersionConstraint{}, fmt.Errorf("while parsing maximum constraint %q: %w", constraint, err)
 		}
 		sentence = fmt.Sprintf("version %s or lower", max)
 	} else {
-		d("constraint type is exact match")
+		D("constraint type is exact match")
 		exact, err := NewFirefoxVersion(constraint)
 		if err != nil {
 			return FirefoxVersionConstraint{}, fmt.Errorf("while parsing exact match constraint %q: %w", constraint, err)
@@ -95,7 +95,7 @@ func (ffv FirefoxVersion) LesserOrEqual(other FirefoxVersion) bool {
 	if other.Minor == -1 || ffv.Minor == -1 {
 		return ffv.Major <= other.Major
 	}
-	d("check that %d <= %d or %d = %d and %d <= %d", ffv.Major, other.Major, ffv.Major, other.Major, ffv.Minor, other.Minor)
+	D("check that %d <= %d or %d = %d and %d <= %d", ffv.Major, other.Major, ffv.Major, other.Major, ffv.Minor, other.Minor)
 	return ffv.Major < other.Major || (ffv.Major == other.Major && ffv.Minor <= other.Minor)
 }
 
@@ -110,8 +110,8 @@ func (ffv FirefoxVersion) String() string {
 
 // FulfilledBy checks if version ∈ [constraint.min, constraint.max]
 func (constraint FirefoxVersionConstraint) FulfilledBy(version FirefoxVersion) bool {
-	d("checking if %s ∈ [%s, %s]", version, constraint.min, constraint.max)
-	return version.GreaterOrEqual(constraint.min) && version.LesserOrEqual(constraint.max)
+	D("checking if %s ∈ [%s, %s]", version, constraint.Min, constraint.Max)
+	return version.GreaterOrEqual(constraint.Min) && version.LesserOrEqual(constraint.Max)
 }
 
 // NewFirefoxVersion turns a version string (90 or 90.0 for example) into a FirefoxVersion.
@@ -124,7 +124,7 @@ func NewFirefoxVersion(stringRepr string, defaultMinor ...string) (FirefoxVersio
 	if len(fragments) == 1 {
 		fragments = append(fragments, defaultMinor...)
 	}
-	d("parsing version %s: fragments is %#v", stringRepr, fragments)
+	D("parsing version %s: fragments is %#v", stringRepr, fragments)
 	major, err := strconv.ParseInt(fragments[0], 10, 64)
 	if err != nil {
 		return FirefoxVersion{}, fmt.Errorf("while converting major segment: %w", err)
