+++ utils_windows_test.go
@@ -1,4 +1,4 @@
-package main
+package ffcss
 
 import (
 	"os"
